Refactoring code and new features (createUser, loginUser)
