--- petStore/src/test/java/karate/request/DeletePet.java
@@ -1,9 +1,0 @@
-package karate.request;
-
-import com.intuit.karate.junit5.Karate;
-public class DeletePet {
-    @Karate.Test
-    Karate DeletePet(){
-        return Karate.run("classpath:karate/features/4deletePet.feature").relativeTo(getClass());
-    }
-}
