--- petStore/src/test/java/karate/request/UpdatePetRunner.java
@@ -1,9 +1,0 @@
-package karate.request;
-
-import com.intuit.karate.junit5.Karate;
-public class UpdatePetRunner {
-    @Karate.Test
-    Karate updatePet(){
-        return Karate.run("classpath:karate/features/3updatePet.feature").relativeTo(getClass());
-    }
-}
