+++ petStore/src/test/java/karate/request/pet/AddPetRunner.java
@@ -1,0 +1,10 @@
+package karate.request.pet;
+
+import com.intuit.karate.junit5.Karate;
+
+public class AddPetRunner {
+    @Karate.Test
+    Karate AddPetTest() {
+        return Karate.run("classpath:karate/features/pet/addPet.feature").relativeTo(getClass());
+    }
+}
