+++ petStore/src/test/java/karate/request/pet/FindPetByIdRunner.java
@@ -1,0 +1,9 @@
+package karate.request.pet;
+
+import com.intuit.karate.junit5.Karate;
+public class FindPetByIdRunner {
+    @Karate.Test
+    Karate FindPetById(){
+        return Karate.run("classpath:karate/features/pet/findPetById.feature").relativeTo(getClass());
+    }
+}
