+++ petStore/src/test/java/karate/request/pet/UpdatePetRunner.java
@@ -1,0 +1,9 @@
+package karate.request.pet;
+
+import com.intuit.karate.junit5.Karate;
+public class UpdatePetRunner {
+    @Karate.Test
+    Karate updatePet(){
+        return Karate.run("classpath:karate/features/pet/updatePet.feature").relativeTo(getClass());
+    }
+}
