+++ petStore/src/test/java/karate/request/user/CreateUserRunner.java
@@ -1,0 +1,9 @@
+package karate.request.user;
+
+import com.intuit.karate.junit5.Karate;
+public class CreateUserRunner {
+    @Karate.Test
+    Karate createUser(){
+        return Karate.run("classpath:karate/features/user/createUser.feature").relativeTo(getClass());
+    }
+}
