+++ petStore/src/test/resources/karate-properties.json
@@ -1,6 +1,10 @@
 {
   "dev": {
     "url": "https://petstore.swagger.io/v2/",
-    "id": 5
+    "id": 5,
+    "message": "5",
+    "status": "pending",
+    "statusUpdate": "sold",
+    "idUser": "2"
   }
 }
