+++ petStore/src/test/resources/karate/features/pet/addPet.feature
@@ -1,3 +1,4 @@
+@addPet
 Feature: Add a new pet
   As QA automation
   I want to add a new pet
@@ -5,9 +6,9 @@ Feature: Add a new pet
 
   Background: Consume the service
     * url url
-    * def requestBody = read('classpath:karate/features/AddPetBody.json')
-    * def responseAddPet = read('classpath:karate/features/AddPetBody.json')
-
+    * def requestBody = read('classpath:karate/features/pet/AddPetBody.json')
+    * def responseAddPet = read('classpath:karate/features/pet/AddPetBody.json')
+  @addPetHappyPath
   Scenario: Create a new pet with valid data
 
     Given path 'pet'
@@ -15,15 +16,17 @@ Feature: Add a new pet
     When method post
     Then status 200
     * match response == responseAddPet
+    And assert response.id == id
+    And assert response.status == status
 
-
+  @addPetNegative
   Scenario Outline: Add a new pet with invalid status
     * def statusAdd = '<status>'
     Given path 'pet'
     And request requestBody
     And status, statusAdd
     When method POST
-    Then status 200 #Arroja un 200, pero en realidad debe ser un 405, esto es un gran bug
+    Then status 200
     * match response == responseAddPet
 
     Examples:
