+++ petStore/src/test/resources/karate/features/pet/deleteResponse.json
@@ -1,0 +1,5 @@
+{
+  "code": 200,
+  "type": "unknown",
+  "message": "5"
+}
