+++ petStore/src/test/resources/karate/features/pet/updatePetBody.json
@@ -14,5 +14,5 @@
       "name": "pet"
     }
   ],
-  "status": "pending"
+  "status": "sold"
 }
