+++ petStore/src/test/resources/karate/features/user/createUserBody.json
@@ -1,0 +1,10 @@
+{
+  "id": "2",
+  "username": "edwinc",
+  "firstName": "Edwin",
+  "lastName": "Cuadro",
+  "email": "cuadro@email.com",
+  "password": "123",
+  "phone": "30303",
+  "userStatus": 0
+}
