+++ petStore/src/test/resources/karate/features/user/createUserResponse.json
@@ -1,0 +1,5 @@
+{
+  "code": "#number",
+  "type": "#string",
+  "message": "#string"
+}
